2009-07-15  Sebastien Pouliot  <sebastien@ximian.com>
	* gdipFunctions.cs: Fix XVisualIDFromVisual to return an IntPtr
	since 64 bits X servers will return a 64 bits value.
	* gdipStructs.cs: Fix XVisualInfo to use IntPtr on fields that
	are 64 bits on 64 bits X servers.
	* Graphics.cs: Adjust CopyFromScreenX11 to work with IntPtr so the
	screenshot code can work on both 32 and 64 bits X servers.
	[Fix bug #521718]


svn path=/trunk/mcs/; revision=137982

+++ mcs/class/System.Drawing/System.Drawing/gdipFunctions.cs
@@ -1515,7 +1515,7 @@ namespace System.Drawing
 		internal extern static IntPtr XGetVisualInfo (IntPtr display, int vinfo_mask, ref XVisualInfo vinfo_template, ref int nitems);
 
 		[DllImport ("libX11", EntryPoint="XVisualIDFromVisual")]
-		internal extern static int XVisualIDFromVisual(IntPtr visual);
+		internal extern static IntPtr XVisualIDFromVisual (IntPtr visual);
 
 		[DllImport ("libX11", EntryPoint="XFree")]
 		internal extern static void XFree (IntPtr data);
